+++ src/routes/playground/tags/playground/tags/playground-output/playground-output.marko
@@ -1,7 +1,7 @@
-import type { WorkspaceFile } from "<workspace>";
+import type { File } from "app/util/workspace";
 
 export interface Input {
-  files: WorkspaceFile[];
+  files: File[];
   selected: number;
 }
