--- src/routes/playground/tags/workspace/utils/to-data-uri.ts
@@ -1,3 +1,0 @@
-export function toDataURI(type: string, code: string) {
-  return `data:${type};charset=utf-8;base64,${btoa(code)}`;
-}
