--- src/routes/playground/tags/workspace/workspace.marko
@@ -1,6 +1,0 @@
-client import { subscribe, type Workspace } from "./utils/compiler";
-client export { update, type Workspace, type WorkspaceFile } from "./utils/compiler";
-
-let/workspace=({} as Partial<Workspace>)
-return=workspace
-script -- subscribe(v => workspace = v, $signal);
