+++ src/util/hasher.ts
