+++ src/util/workspace/minify-script-plugin.ts
